seperate function into files | add minirt_error.h for error text.
